Moved the main part of XMPUtil to jabref.XMPUtilMain and injected a bit more preferences
